Moved cooldowns and errors from discord to discord.commands
